.circleci: Remove all short-circuits that use pipeline.git.base_revision
The author clearly understood <<pipeline.git.base_revison>> to be the
appropriate $(git merge-base) for the PR.  I also understood it that way,
as I remember reviewing the PR that originally added the check.

But that's not quite what it is.  These days at least, it's set to the
"build prior to the one being built"[1].  As observed in one job[2], this
means that if you push to your PR branch after your initial push, it is
then set to the previous tip of your PR branch, and *not* to the
merge-base.

Honestly, I don't know how anyone at CircleCI thought this would be
useful.  It boggles them mind.  IMO, it smells of someone implementing a
customer request without actually understanding what the customer wants or
giving thought to how it might actually be used.

Anyway, that means that the check is so severely broken that it undermines
the confidence we get from CI.  Get rid of these broken optimizations.

[1]: https://circleci.com/docs/2.0/pipeline-variables/
[2]: https://app.circleci.com/pipelines/github/datawire/edge-stack/121/workflows/824fe6b7-283f-4334-b4f1-2e1b1698cbc1/jobs/231

Signed-off-by: Luke Shumaker <lukeshu@datawire.io>