OAuth2: Make sure we actually handle CORS requests for known filter endpoints.
Add support for handling CORS request and CORS preflight request on the known
endpoints for `/.ambassador/oauth2/logout` and `/.ambassador/oauth2/multicooke`.
The appropriate CORS headers are added to the response.

We ensure that the `Access-Control-Allow-Origin` header is set because some
browsers are more strict than others. In certain versions of Safari we have
witnessed not including the origin to cause Safari to reject the CORS
request.

While we respond to both CORS preflights and to CORS requests, we generally do
_not_ allow configuring the response: when the OAuth2 filter is in play,
there's pretty much only one Right Way to Respond.

Co-authored-by: Lance Austin <laustin@datawire.io>
Signed-off-by: Flynn <flynn@datawire.io>