fixes to support apple silicon development (#2949)
* introduce BUILD_ARCH and upgrade `yq` to support Apple Silicon

By default, docker build will run based on your Host OS. This is a problem for developing on Apple Silicon because a lot of our tooling and build process. This updates calls to `docker build` to force using the platform defined in the env variable BUILD_ARCH. It defaults to `linux/amd64` so no changes to the devloop.

`yq` v3 has a transitive dependency that doesn't work with arm64. This moves it from the emissary repository where it is no longer used and upgrades it to v4. v4 introduced new syntax so the arguments were adjusted to match the breaking changes between versions.

Signed-off-by: Lance Austin <laustin@datawire.io>