fix: ensure custom resources round trip properly
This commit renames TestConvert to TestConvertThroughHub to
make it a little more obvious what it is testing. The round-trip
equality checks are moved from a string comparison to a semantic
deep equality comparison due to the fact that not all fields
convert 1-1. Things like adding default AmbassadorID, mangling it
and enums filling default values all caused the existing string
comparisons to fail.

Fixes for FilterPolicy and ExternalFilters based on visibilty from
the failing test were addressed as well.

Signed-off-by: Lance Austin <laustin@datawire.io>